Move Announcements out of device
