+++ src/components/Application.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import { Redirect, Route, RouteComponentProps, Switch } from 'react-router'
 
 import AdminContainer from '../containers/admin'
+import Announcements from '../containers/AnnouncementsContainer'
 import Academy from '../containers/device'
-import Announcements from '../containers/device/AnnouncementsContainer'
 import Game from '../containers/GameContainer'
 import NavigationBar from './NavigationBar'
 import NotFound from './NotFound'
